+++ src/Cache/src/Bootloader/CacheBootloader.php
@@ -11,6 +11,7 @@ use Spiral\Boot\EnvironmentInterface;
 use Spiral\Cache\CacheManager;
 use Spiral\Cache\CacheStorageProviderInterface;
 use Spiral\Cache\Config\CacheConfig;
+use Spiral\Cache\Core\CacheInjector;
 use Spiral\Cache\Storage\ArrayStorage;
 use Spiral\Cache\Storage\FileStorage;
 use Spiral\Config\ConfiguratorInterface;
@@ -24,8 +25,7 @@ final class CacheBootloader extends Bootloader
         CacheManager::class => [self::class, 'initCacheManager'],
     ];
 
-    /** @var ConfiguratorInterface */
-    private $config;
+    private ConfiguratorInterface $config;
 
     public function __construct(ConfiguratorInterface $config)
     {
@@ -44,9 +44,7 @@ final class CacheBootloader extends Bootloader
     {
         $this->initConfig($env, $dirs);
 
-        $container->bind(CacheInterface::class, function (CacheManager $manager) {
-            return $manager->storage();
-        });
+        $container->bindInjector(CacheInterface::class, CacheInjector::class);
     }
 
     /**